back again

+++ rename.py
@@ -4,5 +4,6 @@ path = 'valid'
 
 j = 1
 for i in os.listdir(path):
-    os.rename(path+'/'+i, path+'/'+str(j)+'.png')
+    if not os.path.exists(path+'/'+i):
+        os.rename(path+'/'+i, path+'/'+str(j)+'.png')
     j+=1
